+++ src/main/java/com/project/cheerha/common/scheduler/consumer/TaskConsumer.java
@@ -1,5 +1,8 @@
-package com.project.cheerha.common.scheduler;
+package com.project.cheerha.common.scheduler.consumer;
 
+import com.project.cheerha.common.scheduler.core.InstanceManager;
+import com.project.cheerha.common.scheduler.core.TaskHandler;
+import com.project.cheerha.common.scheduler.repository.TaskRepository;
 import jakarta.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
