+++ src/main/java/com/project/cheerha/common/scheduler/core/CentralScheduler.java
@@ -1,5 +1,6 @@
-package com.project.cheerha.common.scheduler;
+package com.project.cheerha.common.scheduler.core;
 
+import com.project.cheerha.common.scheduler.producer.TaskRegister;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
