+++ src/main/java/com/project/cheerha/common/scheduler/core/TaskHandler.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.common.scheduler;
+package com.project.cheerha.common.scheduler.core;
 
 import java.util.Map;
